requirements.txt
boto3>=1.28.0
gradio>=3.40.0
pandas>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
soundfile>=0.12.1
numpy>=1.24.0
requests>=2.31.0
